+++ driver-core/CHANGELOG.rst
@@ -36,6 +36,7 @@ Merged from 2.0 branch:
   executions (JAVA-723)
 - [bug] Fix buffer leaks when compression is enabled (JAVA-734).
 - [improvement] Use Netty's pooled ByteBufAllocator by default (JAVA-756)
+- [improvement] Expose "unsafe" paging state API (JAVA-759)
 
 
 2.1.5: